Statistical modelling for individual taxa in a phyloseq

tax_model provides a simple framework to statistically model the abundance of individual taxa in your data. You can choose which type of statistical model you want to fit, and you can choose at which rank and (optionally) which specific taxa to fit statistical models for. tax_model takes a phyloseq and returns a list of statistical models, one model for each taxon. The same independent variables are used for all models, as specified in variables or formula argument (latter takes precedence).

taxatree_models runs tax_model on every taxon at multiple taxonomic ranks (you choose which ranks with the plural tax_levels argument), and returns the results as a named nested list designed for use with taxatree_plots. One list per rank, one model per taxon at each rank.

type = "bbdml" will run beta binomial regression model(s) using the corncob package. For bbdml the same formula/variables is/are used for modelling both the abundance and dispersion parameters.

tax_model(
  ps,
  tax_level,
  type = "bbdml",
  variables = NULL,
  formula = NULL,
  taxa = NULL,
  verbose = TRUE,
  ...
)

taxatree_models(
  ps,
  tax_levels = NULL,
  type = "bbdml",
  variables = NULL,
  formula = NULL,
  verbose = TRUE,
  ...
)

Arguments

ps
phyloseq object

tax_level
name of taxonomic rank to aggregate to and model taxa at

type
name of modelling function to use

variables
vector of variable names to use in statistical model as right hand side (ignored if formula given)

formula
(alternative to variables arg) right hand side of a formula, as a formula object or character value

taxa
taxa to model (named, numbered, logical selection, or defaulting to all if NULL)

verbose
message about progress and any taxa name modifications

...
extra args passed directly to modelling function

tax_levels
names of ranks to model taxa at (or their index) or defaults to all ranks except the first

Value

named list of model objects or list of lists

Details

tax_model and taxatree_models can use parallel processing with the future package. This can speed up analysis if you have many taxa to model. Run a line like this beforehand: future::plan(future::multisession, workers = 3)

See also

Examples

# corncob stats
library(corncob)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

data(dietswap, package = "microbiome")
ps <- dietswap

# create some binary variables for easy visualisation
ps <- ps %>% ps_mutate(
  female = if_else(sex == "female", 1, 0, NaN),
  overweight = if_else(bmi_group == "overweight", 1, 0, NaN),
  obese = if_else(bmi_group == "obese", 1, 0, NaN)
)

#> Loading required package: phyloseq

# This example HITChip dataset has some taxa with the same name for phylum and family...
# We can fix problems like this with the tax_prepend_ranks function
ps <- tax_prepend_ranks(ps)

# filter out rare taxa (it is often difficult to fit multivariable models to rare taxa)
ps <- ps %>% tax_filter(min_prevalence = 0.1, min_total_abundance = 10000)

#> Proportional min_prevalence given: 0.1 --> min 23/222 samples.

# specify variables used for modelling
VARS <- c("female", "overweight", "obese")

# Model first 3 genera using all VARS as predictors (just for a quick test)
models <- tax_model(ps, tax_level = "Genus", taxa = 1:3, variables = VARS)

#> Modelling: G: Akkermansia

#> Modelling: G: Allistipes et rel.

#> Modelling: G: Anaerostipes caccae et rel.

# Alternative method using formula arg instead of variables to produce identical results
models2 <-
  tax_model(ps, tax_level = "Genus", taxa = 1:3, formula = ~ female + overweight + obese)

#> Modelling: G: Akkermansia

#> Modelling: G: Allistipes et rel.

#> Modelling: G: Anaerostipes caccae et rel.

all.equal(models, models2) # should be TRUE

#> [1] TRUE

# Model only one genus, NOTE the modified name,
# which was returned by tax_prepend_ranks defaults
models3 <- ps %>%
  tax_model(tax_level = "Genus", taxa = "G: Bacteroides fragilis et rel.", variables = VARS)

#> Modelling: G: Bacteroides fragilis et rel.

# Model all taxa at multiple taxonomic ranks (ranks 1 and 2)
# using only female variable as predictor
models4 <- taxatree_models(ps, tax_levels = 1:2, formula = ~female, verbose = FALSE)

#> 2021-04-01 09:24:47 - modelling at level: Phylum

#> 2021-04-01 09:24:47 - modelling at level: Family

# modelling proportion with simple linear regression is also possible via type = lm
# and transforming the taxa to compositional first
models_lm <- ps %>%
  microbiome::transform("compositional") %>%
  tax_model(tax_level = "Genus", taxa = 1:3, variables = VARS, type = "lm")

#> Modelling: G: Akkermansia

#> Modelling: G: Allistipes et rel.

#> Modelling: G: Anaerostipes caccae et rel.