Beta binomial taxon models with corncob

Takes a phyloseq and returns a list of models, one for each taxon. Same independent variables for all models. Same variables for modelling both the abundance and dispersion parameters. P-values are from wald tests for simplicity and speed... maybe I will change this in the future.

Use this function with the output of model_tax_corncob. models2stats_corncob splits the statistical results extracted from corncob models and groups them by the independent variable name that they refer to. One dataframe of this list can then be joined to the output of taxatree_nodes to prepare for taxonomic heat tree graph visualisation of taxon-variable associations.

model_tax_corncob(ps, tax_level, variables, taxa = "all")

models2stats_corncob(taxon_models)

Arguments

ps
phyloseq object

tax_level
name of taxonomic rank to aggregate to and model taxa at

variables
vector of variable names to use in corncob beta binomial model

taxa
taxa to model (named, numbered, or defaulting to 'all')

taxon_models
named list output of model_tax_corncob

Value

list of dataframes, one df per independent variable

Details

model_tax_corncob can use parallel processing with the future package. This speeds up analysis if you have many taxa to model. Run a line like this beforehand: future::plan(future::multisession, workers = 3)

Examples

# corncob stats testing

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

library(tibble)
library(phyloseq)
library(microbiome)

#> Loading required package: ggplot2

#> 
#> microbiome R package (microbiome.github.com)
#>     
#> 
#> 
#>  Copyright (C) 2011-2020 Leo Lahti, 
#>     Sudarshan Shetty et al. <microbiome.github.io>

#> 
#> Attaching package: 'microbiome'

#> The following object is masked from 'package:ggplot2':
#> 
#>     alpha

#> The following object is masked from 'package:vegan':
#> 
#>     diversity

#> The following object is masked from 'package:base':
#> 
#>     transform

library(corncob)
library(future)
library(future.apply)

data(dietswap)
ps <- dietswap

# create some binary variables for easy visualisation
sample_data(ps)$female <- if_else(sample_data(ps)$sex == "female", 1, 0, NaN)
sample_data(ps)$overweight <- if_else(sample_data(ps)$bmi_group == "overweight", 1, 0, NaN)
sample_data(ps)$obese <- if_else(sample_data(ps)$bmi_group == "obese", 1, 0, NaN)

# This example dataset has some taxa with the same name for phylum and family...
# We can fix problems like this with the tax_prepend_ranks function
ps <- tax_prepend_ranks(ps)
# this example dataset also has no root, this is unusual and needs to be fixed
tax_table(ps) <- cbind(root = "root", tax_table(ps))

# filter out rare taxa
ps <- ps %>% tax_filter(min_prevalence = 0.2, min_total_abundance = 10000)

#> Proportional min_prevalence given: 0.2 --> min 45/222 samples.

# specify variables used for modelling
variables <- c("female", "overweight", "obese")

# model_tax_corncob can use parallel processing with the futures package
# this speeds up analysis if you have many taxa to model
# e.g. # future::plan(future::multisession, workers = 3)

# prep for tree viz #
# model taxa at all taxonomic levels (except root)
tax_models_list <- lapply(
  X = phyloseq::rank_names(ps)[-1],
  function(r) {
    message(Sys.time(), " - modelling at level: ", r)
    models <- model_tax_corncob(ps, tax_level = r, variables = variables, taxa = "all")
    return(models)
  }
)

#> 2021-01-19 11:37:37 - modelling at level: Phylum

#> 2021-01-19 11:37:39 - modelling at level: Family

#> 2021-01-19 11:37:41 - modelling at level: Genus

# Flatten all taxonomic level lists into one list of models
# (this is why taxon names must be completely identifiable across ranks)
flat_models_list <- purrr::flatten(tax_models_list)
var_stats <- models2stats_corncob(flat_models_list)

# future::plan(future::sequential) # to turn parallel processing back off