Replace unknown, NA, or short tax_table values

Identifies phyloseq tax_table values as unknown or uninformative and replaces them with the first informative value from a higher taxonomic rank.

Short values in phyloseq tax_table are typically empty strings or " ", or "g__" etc. so it is helpful to replace them. Set min_length = 0 to avoid filtering on length.

Values in unknowns are also removed, even if longer than min_length. It is up to the user to specify sensible values in unknowns if their dataset has other unwanted values.

NA values are also replaced.

tax_fill_unknowns(
  x,
  min_length = 4,
  unknowns = c("unknown", paste0(c("p", "c", "o", "f", "g", "s"), "__")),
  levels = phyloseq::rank_names(x)
)

Arguments

x
phyloseq or tax_table (taxonomyTable)

min_length
replace strings shorter than this

unknowns
also replace strings matching any in this vector

levels
names of taxonomic levels to modify, defaults to all

Value

object same class as x

Examples

library(dplyr)
library(tibble)
library(phyloseq)
library(microbiome)

data(dietswap)
ps <- dietswap

tt <- data.frame(tax_table(ps))

# create unknowns to test filling
ntax <- ntaxa(ps)
set.seed(123)
tt[sample(1:ntax, 30), 3] <- "g__"
tt[sample(1:ntax, 20), 2] <- "f__"
tt[sample(1:ntax, 10), 1] <- "p__"
tt[sample(1:ntax, 10), 3] <- "unknown"

tt <- tax_table(as.matrix.data.frame(tt))
tax_table(ps) <- tt

# works on phyloseq objects, and the defaults should solve most problems
ps

#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 3 taxonomic ranks ]

tax_fill_unknowns(ps)

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Proteobacteria; Aquabacterium

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Bacteroidetes; Bacteroides splachnicus et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Actinobacteria; g__

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; f__; Clostridium felsineum et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; f__; Clostridium felsineum et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Clostridium cluster IV; Eubacterium siraeum et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Bacilli; g__

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Proteobacteria; Methylobacterium

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Clostridium cluster IV; g__

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Bacteroidetes; Prevotella melaninogenica et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; f__; Uncultured Clostridiales I

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; f__; Uncultured Clostridiales I

#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 3 taxonomic ranks ]

tt %>% head(50)

#> Taxonomy Table:     [50 taxa by 3 taxonomic ranks]:
#>                                    Phylum           
#> Actinomycetaceae                   "Actinobacteria" 
#> Aerococcus                         "Firmicutes"     
#> Aeromonas                          "Proteobacteria" 
#> Akkermansia                        "Verrucomicrobia"
#> Alcaligenes faecalis et rel.       "Proteobacteria" 
#> Allistipes et rel.                 "Bacteroidetes"  
#> Anaerobiospirillum                 "Proteobacteria" 
#> Anaerofustis                       "Firmicutes"     
#> Anaerostipes caccae et rel.        "Firmicutes"     
#> Anaerotruncus colihominis et rel.  "Firmicutes"     
#> Anaerovorax odorimutans et rel.    "Firmicutes"     
#> Aneurinibacillus                   "Firmicutes"     
#> Aquabacterium                      "p__"            
#> Asteroleplasma et rel.             "Firmicutes"     
#> Atopobium                          "Actinobacteria" 
#> Bacillus                           "Firmicutes"     
#> Bacteroides fragilis et rel.       "Bacteroidetes"  
#> Bacteroides intestinalis et rel.   "Bacteroidetes"  
#> Bacteroides ovatus et rel.         "Bacteroidetes"  
#> Bacteroides plebeius et rel.       "Bacteroidetes"  
#> Bacteroides splachnicus et rel.    "p__"            
#> Bacteroides stercoris et rel.      "Bacteroidetes"  
#> Bacteroides uniformis et rel.      "Bacteroidetes"  
#> Bacteroides vulgatus et rel.       "Bacteroidetes"  
#> Bifidobacterium                    "p__"            
#> Bilophila et rel.                  "Proteobacteria" 
#> Brachyspira                        "Spirochaetes"   
#> Bryantella formatexigens et rel.   "Firmicutes"     
#> Bulleidia moorei et rel.           "Firmicutes"     
#> Burkholderia                       "Proteobacteria" 
#> Butyrivibrio crossotus et rel.     "Firmicutes"     
#> Campylobacter                      "Proteobacteria" 
#> Catenibacterium mitsuokai et rel.  "Firmicutes"     
#> Clostridium (sensu stricto)        "Firmicutes"     
#> Clostridium cellulosi et rel.      "Firmicutes"     
#> Clostridium colinum et rel.        "Firmicutes"     
#> Clostridium difficile et rel.      "Firmicutes"     
#> Clostridium felsineum et rel.      "p__"            
#> Clostridium leptum et rel.         "Firmicutes"     
#> Clostridium nexile et rel.         "Firmicutes"     
#> Clostridium orbiscindens et rel.   "Firmicutes"     
#> Clostridium ramosum et rel.        "Firmicutes"     
#> Clostridium sphenoides et rel.     "Firmicutes"     
#> Clostridium stercorarium et rel.   "Firmicutes"     
#> Clostridium symbiosum et rel.      "Firmicutes"     
#> Clostridium thermocellum et rel.   "Firmicutes"     
#> Collinsella                        "Actinobacteria" 
#> Coprobacillus catenaformis et rel. "Firmicutes"     
#> Coprococcus eutactus et rel.       "Firmicutes"     
#> Corynebacterium                    "Actinobacteria" 
#>                                    Family                     
#> Actinomycetaceae                   "Actinobacteria"           
#> Aerococcus                         "Bacilli"                  
#> Aeromonas                          "Proteobacteria"           
#> Akkermansia                        "Verrucomicrobia"          
#> Alcaligenes faecalis et rel.       "Proteobacteria"           
#> Allistipes et rel.                 "Bacteroidetes"            
#> Anaerobiospirillum                 "f__"                      
#> Anaerofustis                       "Clostridium cluster XV"   
#> Anaerostipes caccae et rel.        "f__"                      
#> Anaerotruncus colihominis et rel.  "Clostridium cluster IV"   
#> Anaerovorax odorimutans et rel.    "Clostridium cluster XI"   
#> Aneurinibacillus                   "Bacilli"                  
#> Aquabacterium                      "Proteobacteria"           
#> Asteroleplasma et rel.             "Asteroleplasma"           
#> Atopobium                          "Actinobacteria"           
#> Bacillus                           "Bacilli"                  
#> Bacteroides fragilis et rel.       "Bacteroidetes"            
#> Bacteroides intestinalis et rel.   "Bacteroidetes"            
#> Bacteroides ovatus et rel.         "Bacteroidetes"            
#> Bacteroides plebeius et rel.       "Bacteroidetes"            
#> Bacteroides splachnicus et rel.    "Bacteroidetes"            
#> Bacteroides stercoris et rel.      "Bacteroidetes"            
#> Bacteroides uniformis et rel.      "f__"                      
#> Bacteroides vulgatus et rel.       "Bacteroidetes"            
#> Bifidobacterium                    "Actinobacteria"           
#> Bilophila et rel.                  "Proteobacteria"           
#> Brachyspira                        "f__"                      
#> Bryantella formatexigens et rel.   "Clostridium cluster XIVa" 
#> Bulleidia moorei et rel.           "Clostridium cluster XVI"  
#> Burkholderia                       "Proteobacteria"           
#> Butyrivibrio crossotus et rel.     "Clostridium cluster XIVa" 
#> Campylobacter                      "f__"                      
#> Catenibacterium mitsuokai et rel.  "Clostridium cluster XVII" 
#> Clostridium (sensu stricto)        "f__"                      
#> Clostridium cellulosi et rel.      "Clostridium cluster IV"   
#> Clostridium colinum et rel.        "Clostridium cluster XIVa" 
#> Clostridium difficile et rel.      "Clostridium cluster XI"   
#> Clostridium felsineum et rel.      "f__"                      
#> Clostridium leptum et rel.         "Clostridium cluster IV"   
#> Clostridium nexile et rel.         "Clostridium cluster XIVa" 
#> Clostridium orbiscindens et rel.   "f__"                      
#> Clostridium ramosum et rel.        "Clostridium cluster XVIII"
#> Clostridium sphenoides et rel.     "Clostridium cluster XIVa" 
#> Clostridium stercorarium et rel.   "Clostridium cluster III"  
#> Clostridium symbiosum et rel.      "Clostridium cluster XIVa" 
#> Clostridium thermocellum et rel.   "Clostridium cluster III"  
#> Collinsella                        "Actinobacteria"           
#> Coprobacillus catenaformis et rel. "Clostridium cluster XVIII"
#> Coprococcus eutactus et rel.       "Clostridium cluster XIVa" 
#> Corynebacterium                    "Actinobacteria"           
#>                                    Genus                               
#> Actinomycetaceae                   "Actinomycetaceae"                  
#> Aerococcus                         "Aerococcus"                        
#> Aeromonas                          "Aeromonas"                         
#> Akkermansia                        "Akkermansia"                       
#> Alcaligenes faecalis et rel.       "Alcaligenes faecalis et rel."      
#> Allistipes et rel.                 "unknown"                           
#> Anaerobiospirillum                 "g__"                               
#> Anaerofustis                       "Anaerofustis"                      
#> Anaerostipes caccae et rel.        "g__"                               
#> Anaerotruncus colihominis et rel.  "Anaerotruncus colihominis et rel." 
#> Anaerovorax odorimutans et rel.    "Anaerovorax odorimutans et rel."   
#> Aneurinibacillus                   "Aneurinibacillus"                  
#> Aquabacterium                      "Aquabacterium"                     
#> Asteroleplasma et rel.             "g__"                               
#> Atopobium                          "Atopobium"                         
#> Bacillus                           "unknown"                           
#> Bacteroides fragilis et rel.       "Bacteroides fragilis et rel."      
#> Bacteroides intestinalis et rel.   "Bacteroides intestinalis et rel."  
#> Bacteroides ovatus et rel.         "Bacteroides ovatus et rel."        
#> Bacteroides plebeius et rel.       "Bacteroides plebeius et rel."      
#> Bacteroides splachnicus et rel.    "Bacteroides splachnicus et rel."   
#> Bacteroides stercoris et rel.      "Bacteroides stercoris et rel."     
#> Bacteroides uniformis et rel.      "Bacteroides uniformis et rel."     
#> Bacteroides vulgatus et rel.       "Bacteroides vulgatus et rel."      
#> Bifidobacterium                    "g__"                               
#> Bilophila et rel.                  "g__"                               
#> Brachyspira                        "Brachyspira"                       
#> Bryantella formatexigens et rel.   "Bryantella formatexigens et rel."  
#> Bulleidia moorei et rel.           "Bulleidia moorei et rel."          
#> Burkholderia                       "Burkholderia"                      
#> Butyrivibrio crossotus et rel.     "Butyrivibrio crossotus et rel."    
#> Campylobacter                      "Campylobacter"                     
#> Catenibacterium mitsuokai et rel.  "Catenibacterium mitsuokai et rel." 
#> Clostridium (sensu stricto)        "Clostridium (sensu stricto)"       
#> Clostridium cellulosi et rel.      "Clostridium cellulosi et rel."     
#> Clostridium colinum et rel.        "g__"                               
#> Clostridium difficile et rel.      "Clostridium difficile et rel."     
#> Clostridium felsineum et rel.      "Clostridium felsineum et rel."     
#> Clostridium leptum et rel.         "Clostridium leptum et rel."        
#> Clostridium nexile et rel.         "Clostridium nexile et rel."        
#> Clostridium orbiscindens et rel.   "unknown"                           
#> Clostridium ramosum et rel.        "g__"                               
#> Clostridium sphenoides et rel.     "g__"                               
#> Clostridium stercorarium et rel.   "Clostridium stercorarium et rel."  
#> Clostridium symbiosum et rel.      "Clostridium symbiosum et rel."     
#> Clostridium thermocellum et rel.   "Clostridium thermocellum et rel."  
#> Collinsella                        "Collinsella"                       
#> Coprobacillus catenaformis et rel. "Coprobacillus catenaformis et rel."
#> Coprococcus eutactus et rel.       "Coprococcus eutactus et rel."      
#> Corynebacterium                    "g__"                               

# this will replace "unknown"s as well as short values including "g__" and "f__"
tax_fill_unknowns(tt) %>% head(50)

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Proteobacteria; Aquabacterium

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Bacteroidetes; Bacteroides splachnicus et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Actinobacteria; g__

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; f__; Clostridium felsineum et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; f__; Clostridium felsineum et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Clostridium cluster IV; Eubacterium siraeum et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Bacilli; g__

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Proteobacteria; Methylobacterium

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Clostridium cluster IV; g__

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; Bacteroidetes; Prevotella melaninogenica et rel.

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; f__; Uncultured Clostridiales I

#> Warning: No non-unknown values to the left of 1+ entries in this row, returning those unchanged: p__; f__; Uncultured Clostridiales I

#> Taxonomy Table:     [50 taxa by 3 taxonomic ranks]:
#>                                    Phylum           
#> Actinomycetaceae                   "Actinobacteria" 
#> Aerococcus                         "Firmicutes"     
#> Aeromonas                          "Proteobacteria" 
#> Akkermansia                        "Verrucomicrobia"
#> Alcaligenes faecalis et rel.       "Proteobacteria" 
#> Allistipes et rel.                 "Bacteroidetes"  
#> Anaerobiospirillum                 "Proteobacteria" 
#> Anaerofustis                       "Firmicutes"     
#> Anaerostipes caccae et rel.        "Firmicutes"     
#> Anaerotruncus colihominis et rel.  "Firmicutes"     
#> Anaerovorax odorimutans et rel.    "Firmicutes"     
#> Aneurinibacillus                   "Firmicutes"     
#> Aquabacterium                      "p__"            
#> Asteroleplasma et rel.             "Firmicutes"     
#> Atopobium                          "Actinobacteria" 
#> Bacillus                           "Firmicutes"     
#> Bacteroides fragilis et rel.       "Bacteroidetes"  
#> Bacteroides intestinalis et rel.   "Bacteroidetes"  
#> Bacteroides ovatus et rel.         "Bacteroidetes"  
#> Bacteroides plebeius et rel.       "Bacteroidetes"  
#> Bacteroides splachnicus et rel.    "p__"            
#> Bacteroides stercoris et rel.      "Bacteroidetes"  
#> Bacteroides uniformis et rel.      "Bacteroidetes"  
#> Bacteroides vulgatus et rel.       "Bacteroidetes"  
#> Bifidobacterium                    "p__"            
#> Bilophila et rel.                  "Proteobacteria" 
#> Brachyspira                        "Spirochaetes"   
#> Bryantella formatexigens et rel.   "Firmicutes"     
#> Bulleidia moorei et rel.           "Firmicutes"     
#> Burkholderia                       "Proteobacteria" 
#> Butyrivibrio crossotus et rel.     "Firmicutes"     
#> Campylobacter                      "Proteobacteria" 
#> Catenibacterium mitsuokai et rel.  "Firmicutes"     
#> Clostridium (sensu stricto)        "Firmicutes"     
#> Clostridium cellulosi et rel.      "Firmicutes"     
#> Clostridium colinum et rel.        "Firmicutes"     
#> Clostridium difficile et rel.      "Firmicutes"     
#> Clostridium felsineum et rel.      "p__"            
#> Clostridium leptum et rel.         "Firmicutes"     
#> Clostridium nexile et rel.         "Firmicutes"     
#> Clostridium orbiscindens et rel.   "Firmicutes"     
#> Clostridium ramosum et rel.        "Firmicutes"     
#> Clostridium sphenoides et rel.     "Firmicutes"     
#> Clostridium stercorarium et rel.   "Firmicutes"     
#> Clostridium symbiosum et rel.      "Firmicutes"     
#> Clostridium thermocellum et rel.   "Firmicutes"     
#> Collinsella                        "Actinobacteria" 
#> Coprobacillus catenaformis et rel. "Firmicutes"     
#> Coprococcus eutactus et rel.       "Firmicutes"     
#> Corynebacterium                    "Actinobacteria" 
#>                                    Family                     
#> Actinomycetaceae                   "Actinobacteria"           
#> Aerococcus                         "Bacilli"                  
#> Aeromonas                          "Proteobacteria"           
#> Akkermansia                        "Verrucomicrobia"          
#> Alcaligenes faecalis et rel.       "Proteobacteria"           
#> Allistipes et rel.                 "Bacteroidetes"            
#> Anaerobiospirillum                 "Proteobacteria.Family"    
#> Anaerofustis                       "Clostridium cluster XV"   
#> Anaerostipes caccae et rel.        "Firmicutes.Family"        
#> Anaerotruncus colihominis et rel.  "Clostridium cluster IV"   
#> Anaerovorax odorimutans et rel.    "Clostridium cluster XI"   
#> Aneurinibacillus                   "Bacilli"                  
#> Aquabacterium                      "Proteobacteria"           
#> Asteroleplasma et rel.             "Asteroleplasma"           
#> Atopobium                          "Actinobacteria"           
#> Bacillus                           "Bacilli"                  
#> Bacteroides fragilis et rel.       "Bacteroidetes"            
#> Bacteroides intestinalis et rel.   "Bacteroidetes"            
#> Bacteroides ovatus et rel.         "Bacteroidetes"            
#> Bacteroides plebeius et rel.       "Bacteroidetes"            
#> Bacteroides splachnicus et rel.    "Bacteroidetes"            
#> Bacteroides stercoris et rel.      "Bacteroidetes"            
#> Bacteroides uniformis et rel.      "Bacteroidetes.Family"     
#> Bacteroides vulgatus et rel.       "Bacteroidetes"            
#> Bifidobacterium                    "Actinobacteria"           
#> Bilophila et rel.                  "Proteobacteria"           
#> Brachyspira                        "Spirochaetes.Family"      
#> Bryantella formatexigens et rel.   "Clostridium cluster XIVa" 
#> Bulleidia moorei et rel.           "Clostridium cluster XVI"  
#> Burkholderia                       "Proteobacteria"           
#> Butyrivibrio crossotus et rel.     "Clostridium cluster XIVa" 
#> Campylobacter                      "Proteobacteria.Family"    
#> Catenibacterium mitsuokai et rel.  "Clostridium cluster XVII" 
#> Clostridium (sensu stricto)        "Firmicutes.Family"        
#> Clostridium cellulosi et rel.      "Clostridium cluster IV"   
#> Clostridium colinum et rel.        "Clostridium cluster XIVa" 
#> Clostridium difficile et rel.      "Clostridium cluster XI"   
#> Clostridium felsineum et rel.      "f__"                      
#> Clostridium leptum et rel.         "Clostridium cluster IV"   
#> Clostridium nexile et rel.         "Clostridium cluster XIVa" 
#> Clostridium orbiscindens et rel.   "Firmicutes.Family"        
#> Clostridium ramosum et rel.        "Clostridium cluster XVIII"
#> Clostridium sphenoides et rel.     "Clostridium cluster XIVa" 
#> Clostridium stercorarium et rel.   "Clostridium cluster III"  
#> Clostridium symbiosum et rel.      "Clostridium cluster XIVa" 
#> Clostridium thermocellum et rel.   "Clostridium cluster III"  
#> Collinsella                        "Actinobacteria"           
#> Coprobacillus catenaformis et rel. "Clostridium cluster XVIII"
#> Coprococcus eutactus et rel.       "Clostridium cluster XIVa" 
#> Corynebacterium                    "Actinobacteria"           
#>                                    Genus                               
#> Actinomycetaceae                   "Actinomycetaceae"                  
#> Aerococcus                         "Aerococcus"                        
#> Aeromonas                          "Aeromonas"                         
#> Akkermansia                        "Akkermansia"                       
#> Alcaligenes faecalis et rel.       "Alcaligenes faecalis et rel."      
#> Allistipes et rel.                 "Bacteroidetes.Genus"               
#> Anaerobiospirillum                 "Proteobacteria.Genus"              
#> Anaerofustis                       "Anaerofustis"                      
#> Anaerostipes caccae et rel.        "Firmicutes.Genus"                  
#> Anaerotruncus colihominis et rel.  "Anaerotruncus colihominis et rel." 
#> Anaerovorax odorimutans et rel.    "Anaerovorax odorimutans et rel."   
#> Aneurinibacillus                   "Aneurinibacillus"                  
#> Aquabacterium                      "Aquabacterium"                     
#> Asteroleplasma et rel.             "Asteroleplasma.Genus"              
#> Atopobium                          "Atopobium"                         
#> Bacillus                           "Bacilli.Genus"                     
#> Bacteroides fragilis et rel.       "Bacteroides fragilis et rel."      
#> Bacteroides intestinalis et rel.   "Bacteroides intestinalis et rel."  
#> Bacteroides ovatus et rel.         "Bacteroides ovatus et rel."        
#> Bacteroides plebeius et rel.       "Bacteroides plebeius et rel."      
#> Bacteroides splachnicus et rel.    "Bacteroides splachnicus et rel."   
#> Bacteroides stercoris et rel.      "Bacteroides stercoris et rel."     
#> Bacteroides uniformis et rel.      "Bacteroides uniformis et rel."     
#> Bacteroides vulgatus et rel.       "Bacteroides vulgatus et rel."      
#> Bifidobacterium                    "Actinobacteria.Genus"              
#> Bilophila et rel.                  "Proteobacteria.Genus"              
#> Brachyspira                        "Brachyspira"                       
#> Bryantella formatexigens et rel.   "Bryantella formatexigens et rel."  
#> Bulleidia moorei et rel.           "Bulleidia moorei et rel."          
#> Burkholderia                       "Burkholderia"                      
#> Butyrivibrio crossotus et rel.     "Butyrivibrio crossotus et rel."    
#> Campylobacter                      "Campylobacter"                     
#> Catenibacterium mitsuokai et rel.  "Catenibacterium mitsuokai et rel." 
#> Clostridium (sensu stricto)        "Clostridium (sensu stricto)"       
#> Clostridium cellulosi et rel.      "Clostridium cellulosi et rel."     
#> Clostridium colinum et rel.        "Clostridium cluster XIVa.Genus"    
#> Clostridium difficile et rel.      "Clostridium difficile et rel."     
#> Clostridium felsineum et rel.      "Clostridium felsineum et rel."     
#> Clostridium leptum et rel.         "Clostridium leptum et rel."        
#> Clostridium nexile et rel.         "Clostridium nexile et rel."        
#> Clostridium orbiscindens et rel.   "Firmicutes.Genus"                  
#> Clostridium ramosum et rel.        "Clostridium cluster XVIII.Genus"   
#> Clostridium sphenoides et rel.     "Clostridium cluster XIVa.Genus"    
#> Clostridium stercorarium et rel.   "Clostridium stercorarium et rel."  
#> Clostridium symbiosum et rel.      "Clostridium symbiosum et rel."     
#> Clostridium thermocellum et rel.   "Clostridium thermocellum et rel."  
#> Collinsella                        "Collinsella"                       
#> Coprobacillus catenaformis et rel. "Coprobacillus catenaformis et rel."
#> Coprococcus eutactus et rel.       "Coprococcus eutactus et rel."      
#> Corynebacterium                    "Actinobacteria.Genus"              

# this will only replace values in Genus column, and will not replace "unknown"s
tax_fill_unknowns(tt, unknowns = NULL, levels = "Genus") %>% head(50)

#> Taxonomy Table:     [50 taxa by 3 taxonomic ranks]:
#>                                    Phylum           
#> Actinomycetaceae                   "Actinobacteria" 
#> Aerococcus                         "Firmicutes"     
#> Aeromonas                          "Proteobacteria" 
#> Akkermansia                        "Verrucomicrobia"
#> Alcaligenes faecalis et rel.       "Proteobacteria" 
#> Allistipes et rel.                 "Bacteroidetes"  
#> Anaerobiospirillum                 "Proteobacteria" 
#> Anaerofustis                       "Firmicutes"     
#> Anaerostipes caccae et rel.        "Firmicutes"     
#> Anaerotruncus colihominis et rel.  "Firmicutes"     
#> Anaerovorax odorimutans et rel.    "Firmicutes"     
#> Aneurinibacillus                   "Firmicutes"     
#> Aquabacterium                      "p__"            
#> Asteroleplasma et rel.             "Firmicutes"     
#> Atopobium                          "Actinobacteria" 
#> Bacillus                           "Firmicutes"     
#> Bacteroides fragilis et rel.       "Bacteroidetes"  
#> Bacteroides intestinalis et rel.   "Bacteroidetes"  
#> Bacteroides ovatus et rel.         "Bacteroidetes"  
#> Bacteroides plebeius et rel.       "Bacteroidetes"  
#> Bacteroides splachnicus et rel.    "p__"            
#> Bacteroides stercoris et rel.      "Bacteroidetes"  
#> Bacteroides uniformis et rel.      "Bacteroidetes"  
#> Bacteroides vulgatus et rel.       "Bacteroidetes"  
#> Bifidobacterium                    "p__"            
#> Bilophila et rel.                  "Proteobacteria" 
#> Brachyspira                        "Spirochaetes"   
#> Bryantella formatexigens et rel.   "Firmicutes"     
#> Bulleidia moorei et rel.           "Firmicutes"     
#> Burkholderia                       "Proteobacteria" 
#> Butyrivibrio crossotus et rel.     "Firmicutes"     
#> Campylobacter                      "Proteobacteria" 
#> Catenibacterium mitsuokai et rel.  "Firmicutes"     
#> Clostridium (sensu stricto)        "Firmicutes"     
#> Clostridium cellulosi et rel.      "Firmicutes"     
#> Clostridium colinum et rel.        "Firmicutes"     
#> Clostridium difficile et rel.      "Firmicutes"     
#> Clostridium felsineum et rel.      "p__"            
#> Clostridium leptum et rel.         "Firmicutes"     
#> Clostridium nexile et rel.         "Firmicutes"     
#> Clostridium orbiscindens et rel.   "Firmicutes"     
#> Clostridium ramosum et rel.        "Firmicutes"     
#> Clostridium sphenoides et rel.     "Firmicutes"     
#> Clostridium stercorarium et rel.   "Firmicutes"     
#> Clostridium symbiosum et rel.      "Firmicutes"     
#> Clostridium thermocellum et rel.   "Firmicutes"     
#> Collinsella                        "Actinobacteria" 
#> Coprobacillus catenaformis et rel. "Firmicutes"     
#> Coprococcus eutactus et rel.       "Firmicutes"     
#> Corynebacterium                    "Actinobacteria" 
#>                                    Family                     
#> Actinomycetaceae                   "Actinobacteria"           
#> Aerococcus                         "Bacilli"                  
#> Aeromonas                          "Proteobacteria"           
#> Akkermansia                        "Verrucomicrobia"          
#> Alcaligenes faecalis et rel.       "Proteobacteria"           
#> Allistipes et rel.                 "Bacteroidetes"            
#> Anaerobiospirillum                 "f__"                      
#> Anaerofustis                       "Clostridium cluster XV"   
#> Anaerostipes caccae et rel.        "f__"                      
#> Anaerotruncus colihominis et rel.  "Clostridium cluster IV"   
#> Anaerovorax odorimutans et rel.    "Clostridium cluster XI"   
#> Aneurinibacillus                   "Bacilli"                  
#> Aquabacterium                      "Proteobacteria"           
#> Asteroleplasma et rel.             "Asteroleplasma"           
#> Atopobium                          "Actinobacteria"           
#> Bacillus                           "Bacilli"                  
#> Bacteroides fragilis et rel.       "Bacteroidetes"            
#> Bacteroides intestinalis et rel.   "Bacteroidetes"            
#> Bacteroides ovatus et rel.         "Bacteroidetes"            
#> Bacteroides plebeius et rel.       "Bacteroidetes"            
#> Bacteroides splachnicus et rel.    "Bacteroidetes"            
#> Bacteroides stercoris et rel.      "Bacteroidetes"            
#> Bacteroides uniformis et rel.      "f__"                      
#> Bacteroides vulgatus et rel.       "Bacteroidetes"            
#> Bifidobacterium                    "Actinobacteria"           
#> Bilophila et rel.                  "Proteobacteria"           
#> Brachyspira                        "f__"                      
#> Bryantella formatexigens et rel.   "Clostridium cluster XIVa" 
#> Bulleidia moorei et rel.           "Clostridium cluster XVI"  
#> Burkholderia                       "Proteobacteria"           
#> Butyrivibrio crossotus et rel.     "Clostridium cluster XIVa" 
#> Campylobacter                      "f__"                      
#> Catenibacterium mitsuokai et rel.  "Clostridium cluster XVII" 
#> Clostridium (sensu stricto)        "f__"                      
#> Clostridium cellulosi et rel.      "Clostridium cluster IV"   
#> Clostridium colinum et rel.        "Clostridium cluster XIVa" 
#> Clostridium difficile et rel.      "Clostridium cluster XI"   
#> Clostridium felsineum et rel.      "f__"                      
#> Clostridium leptum et rel.         "Clostridium cluster IV"   
#> Clostridium nexile et rel.         "Clostridium cluster XIVa" 
#> Clostridium orbiscindens et rel.   "f__"                      
#> Clostridium ramosum et rel.        "Clostridium cluster XVIII"
#> Clostridium sphenoides et rel.     "Clostridium cluster XIVa" 
#> Clostridium stercorarium et rel.   "Clostridium cluster III"  
#> Clostridium symbiosum et rel.      "Clostridium cluster XIVa" 
#> Clostridium thermocellum et rel.   "Clostridium cluster III"  
#> Collinsella                        "Actinobacteria"           
#> Coprobacillus catenaformis et rel. "Clostridium cluster XVIII"
#> Coprococcus eutactus et rel.       "Clostridium cluster XIVa" 
#> Corynebacterium                    "Actinobacteria"           
#>                                    Genus                               
#> Actinomycetaceae                   "Actinomycetaceae"                  
#> Aerococcus                         "Aerococcus"                        
#> Aeromonas                          "Aeromonas"                         
#> Akkermansia                        "Akkermansia"                       
#> Alcaligenes faecalis et rel.       "Alcaligenes faecalis et rel."      
#> Allistipes et rel.                 "unknown"                           
#> Anaerobiospirillum                 "Proteobacteria.Genus"              
#> Anaerofustis                       "Anaerofustis"                      
#> Anaerostipes caccae et rel.        "Firmicutes.Genus"                  
#> Anaerotruncus colihominis et rel.  "Anaerotruncus colihominis et rel." 
#> Anaerovorax odorimutans et rel.    "Anaerovorax odorimutans et rel."   
#> Aneurinibacillus                   "Aneurinibacillus"                  
#> Aquabacterium                      "Aquabacterium"                     
#> Asteroleplasma et rel.             "Asteroleplasma.Genus"              
#> Atopobium                          "Atopobium"                         
#> Bacillus                           "unknown"                           
#> Bacteroides fragilis et rel.       "Bacteroides fragilis et rel."      
#> Bacteroides intestinalis et rel.   "Bacteroides intestinalis et rel."  
#> Bacteroides ovatus et rel.         "Bacteroides ovatus et rel."        
#> Bacteroides plebeius et rel.       "Bacteroides plebeius et rel."      
#> Bacteroides splachnicus et rel.    "Bacteroides splachnicus et rel."   
#> Bacteroides stercoris et rel.      "Bacteroides stercoris et rel."     
#> Bacteroides uniformis et rel.      "Bacteroides uniformis et rel."     
#> Bacteroides vulgatus et rel.       "Bacteroides vulgatus et rel."      
#> Bifidobacterium                    "Actinobacteria.Genus"              
#> Bilophila et rel.                  "Proteobacteria.Genus"              
#> Brachyspira                        "Brachyspira"                       
#> Bryantella formatexigens et rel.   "Bryantella formatexigens et rel."  
#> Bulleidia moorei et rel.           "Bulleidia moorei et rel."          
#> Burkholderia                       "Burkholderia"                      
#> Butyrivibrio crossotus et rel.     "Butyrivibrio crossotus et rel."    
#> Campylobacter                      "Campylobacter"                     
#> Catenibacterium mitsuokai et rel.  "Catenibacterium mitsuokai et rel." 
#> Clostridium (sensu stricto)        "Clostridium (sensu stricto)"       
#> Clostridium cellulosi et rel.      "Clostridium cellulosi et rel."     
#> Clostridium colinum et rel.        "Clostridium cluster XIVa.Genus"    
#> Clostridium difficile et rel.      "Clostridium difficile et rel."     
#> Clostridium felsineum et rel.      "Clostridium felsineum et rel."     
#> Clostridium leptum et rel.         "Clostridium leptum et rel."        
#> Clostridium nexile et rel.         "Clostridium nexile et rel."        
#> Clostridium orbiscindens et rel.   "unknown"                           
#> Clostridium ramosum et rel.        "Clostridium cluster XVIII.Genus"   
#> Clostridium sphenoides et rel.     "Clostridium cluster XIVa.Genus"    
#> Clostridium stercorarium et rel.   "Clostridium stercorarium et rel."  
#> Clostridium symbiosum et rel.      "Clostridium symbiosum et rel."     
#> Clostridium thermocellum et rel.   "Clostridium thermocellum et rel."  
#> Collinsella                        "Collinsella"                       
#> Coprobacillus catenaformis et rel. "Coprobacillus catenaformis et rel."
#> Coprococcus eutactus et rel.       "Coprococcus eutactus et rel."      
#> Corynebacterium                    "Actinobacteria.Genus"              

# larger example tax_table shows 1000s rows still fast, from microbiomeutilities package
# library(microbiomeutilities)
# data("hmp2")
# tax_fill_unknowns(hmp2)