Fixing duplication problems in phyloseq tax_table

tax_fill_unknowns replaces any specified unknown names with a fully unique name made by pasting together all the higher taxonomic levels and the current rank

taxon_fill_unknowns is an internal helper function for tax_fill_unknowns (so don't use this directly!)

tax_prepend_ranks is a helper function to prepend the start of rank name to each taxon at each rank (useful particularly in case of duplicated taxa names across ranks, e.g. dietswap dataset)

tax_fill_unknowns(ps, unknowns, uniq_from = 1)

taxon_fill_unknowns(taxon_row, unknowns, uniq_from)

tax_prepend_ranks(ps, sep = ": ", nchar = 1)

Arguments

ps
phyloseq object

unknowns
character vector of names that should be 'fixed' e.g. c("", "unknown")

uniq_from
the taxonomic rank from which the new filled name should become unique

taxon_row
a row of a phyloseq tax_table object

sep
for tax_prepend_ranks (what to paste in between rank initial and taxon name)

nchar
number of characters to use from rank_names

Value

a modified phyloseq (or a 1-row matrix for taxon_fill_unknowns)

Details

tax_fill_unknowns supports parallel processing with future package, which might be useful if tax_table(ps) is massive (1000s rows) e.g. future::plan("future::multisession", workers = 3)

Examples

library(dplyr)
library(microbiome)
data(dietswap)
ps <- dietswap

phyloseq::tax_table(ps) %>% head()

#> Taxonomy Table:     [6 taxa by 3 taxonomic ranks]:
#>                              Phylum            Family           
#> Actinomycetaceae             "Actinobacteria"  "Actinobacteria" 
#> Aerococcus                   "Firmicutes"      "Bacilli"        
#> Aeromonas                    "Proteobacteria"  "Proteobacteria" 
#> Akkermansia                  "Verrucomicrobia" "Verrucomicrobia"
#> Alcaligenes faecalis et rel. "Proteobacteria"  "Proteobacteria" 
#> Allistipes et rel.           "Bacteroidetes"   "Bacteroidetes"  
#>                              Genus                         
#> Actinomycetaceae             "Actinomycetaceae"            
#> Aerococcus                   "Aerococcus"                  
#> Aeromonas                    "Aeromonas"                   
#> Akkermansia                  "Akkermansia"                 
#> Alcaligenes faecalis et rel. "Alcaligenes faecalis et rel."
#> Allistipes et rel.           "Allistipes et rel."          

# test non-unique genera problem handling
phyloseq::tax_table(ps)[c(3, 4, 56), "Genus"] <- "unknown"
phyloseq::tax_table(ps) %>% head()

#> Taxonomy Table:     [6 taxa by 3 taxonomic ranks]:
#>                              Phylum            Family           
#> Actinomycetaceae             "Actinobacteria"  "Actinobacteria" 
#> Aerococcus                   "Firmicutes"      "Bacilli"        
#> Aeromonas                    "Proteobacteria"  "Proteobacteria" 
#> Akkermansia                  "Verrucomicrobia" "Verrucomicrobia"
#> Alcaligenes faecalis et rel. "Proteobacteria"  "Proteobacteria" 
#> Allistipes et rel.           "Bacteroidetes"   "Bacteroidetes"  
#>                              Genus                         
#> Actinomycetaceae             "Actinomycetaceae"            
#> Aerococcus                   "Aerococcus"                  
#> Aeromonas                    "unknown"                     
#> Akkermansia                  "unknown"                     
#> Alcaligenes faecalis et rel. "Alcaligenes faecalis et rel."
#> Allistipes et rel.           "Allistipes et rel."          

# supports parallel processing with future package
# useful if tax_table(ps) is massive (1000s rows)
# e.g. future::plan("future::multisession", workers = 3)

# setting the uniq_from argument allows us to specify that the "unknown" genera
# can be uniquely identified using the 2nd taxonomic rank, which is Family in this example
# (and any ranks below that, which would be concatenated, e.g. try the default uniq_from = 1)

ps <- tax_fill_unknowns(ps, unknowns = "unknown", uniq_from = 2)

# don't forget to close any parallel sessions
# future::plan("future::sequential")

phyloseq::tax_table(ps) %>% head()

#> Taxonomy Table:     [6 taxa by 3 taxonomic ranks]:
#>                              Phylum            Family           
#> Actinomycetaceae             "Actinobacteria"  "Actinobacteria" 
#> Aerococcus                   "Firmicutes"      "Bacilli"        
#> Aeromonas                    "Proteobacteria"  "Proteobacteria" 
#> Akkermansia                  "Verrucomicrobia" "Verrucomicrobia"
#> Alcaligenes faecalis et rel. "Proteobacteria"  "Proteobacteria" 
#> Allistipes et rel.           "Bacteroidetes"   "Bacteroidetes"  
#>                              Genus                         
#> Actinomycetaceae             "Actinomycetaceae"            
#> Aerococcus                   "Aerococcus"                  
#> Aeromonas                    "Proteobacteria.Genus"        
#> Akkermansia                  "Verrucomicrobia.Genus"       
#> Alcaligenes faecalis et rel. "Alcaligenes faecalis et rel."
#> Allistipes et rel.           "Allistipes et rel."          

ps %>%
  tax_prepend_ranks() %>%
  tax_table() %>%
  head()

#> Taxonomy Table:     [6 taxa by 3 taxonomic ranks]:
#>                              Phylum               Family              
#> Actinomycetaceae             "P: Actinobacteria"  "F: Actinobacteria" 
#> Aerococcus                   "P: Firmicutes"      "F: Bacilli"        
#> Aeromonas                    "P: Proteobacteria"  "F: Proteobacteria" 
#> Akkermansia                  "P: Verrucomicrobia" "F: Verrucomicrobia"
#> Alcaligenes faecalis et rel. "P: Proteobacteria"  "F: Proteobacteria" 
#> Allistipes et rel.           "P: Bacteroidetes"   "F: Bacteroidetes"  
#>                              Genus                            
#> Actinomycetaceae             "G: Actinomycetaceae"            
#> Aerococcus                   "G: Aerococcus"                  
#> Aeromonas                    "G: Proteobacteria.Genus"        
#> Akkermansia                  "G: Verrucomicrobia.Genus"       
#> Alcaligenes faecalis et rel. "G: Alcaligenes faecalis et rel."
#> Allistipes et rel.           "G: Allistipes et rel."          

# end examples