Plot taxonomic stats on a ggraph tree

Uses ggraph. See tree-development vignette.

taxatree_plot(
  graph_object,
  layout_coords,
  size_stat = "taxon_mean",
  colour_stat = "b_mu",
  max_node_size = 8,
  max_edge_width = max_node_size - 2,
  sig_var = "p_mu",
  sig_threshold = 0.05,
  sig_stroke_width = 1.5,
  colour_palette = "Green-Brown",
  reverse_colours = FALSE,
  luminance_l2 = 80,
  colour_lims = c(-2, 2)
)

Arguments

graph_object
tidygraph made with taxatree_nodes taxatree_edges

layout_coords
dataframe of x and y coordinates

size_stat
name of variable to scale size of nodes and edges (natural log scale)

colour_stat
name of variable to scale colour/fill of nodes and edges (natural log scale)

max_node_size
set this to avoid overlapping nodes at different scales

max_edge_width
max edge width (optionally relative to max node size)

sig_var
name of variable used to define statistical significance (outlines significant nodes) - NULL for none

sig_threshold
value of sig_var below which statistical significance outlines are drawn

sig_stroke_width
fixed width of statistical significance node outlines

colour_palette
any diverging hcl colour palette name see: colorspace::hcl_palettes("diverging", n = 7, plot = TRUE)

reverse_colours
reverse direction of colour scale?

luminance_l2
luminance of colourscale midpoint (higher is lighter)

colour_lims
limits of colour and fill scale, used to harmonise across several plots (values squished to fit this range!)

Value

ggraph ggplot

Examples

#