Draw labelled key to accompany taxatree_plots

Draw labelled key to accompany taxatree_plots

taxatree_plotkey(
  data,
  ...,
  size_stat = list(prevalence = prev),
  node_size_range = c(1.5, 5),
  edge_width_range = node_size_range * 0.8,
  size_guide = "none",
  size_trans = "identity",
  colour = "lightgrey",
  edge_alpha = 0.7,
  title = "Key",
  title_size = 14,
  taxon_renamer = identity,
  .combine_label = any,
  .draw_label = TRUE,
  .calc_label = TRUE,
  layout = "tree",
  layout_seed = NA,
  circular = identical(layout, "tree"),
  node_sort = NULL,
  add_circles = isTRUE(circular),
  drop_ranks = TRUE
)

Arguments

data
ps_extra (or phyloseq)

...
logical conditions for labelling e.g. rank == "Phylum", p.value < 0.1 | taxon %in% listOfTaxa

size_stat
named list of length 1, giving function calculated for each taxon, to determine the size of nodes (and edges). Name used as size legend title.

node_size_range
min and max node sizes, decrease to avoid node overlap

edge_width_range
min and max edge widths

size_guide
guide for node sizes, try "none", "legend" or ggplot2::guide_legend()

size_trans
transformation for size scale you can use (the name of) any transformer from the scales package, such as "identity", "log1p", or "sqrt"

colour
fixed colour and fill of nodes and edges

edge_alpha
fixed alpha value for edges

title
title of plot (NULL for none)

title_size
font size of title

taxon_renamer
function that takes taxon names and returns modified names for labels

.combine_label
all or any: function to combine multiple logical "label" values for a taxon (relevant if taxatree_stats already present in data)

.draw_label
should labels be drawn, or just the bare tree, set this to FALSE if you want to draw customised labels with taxatree_plot_labels afterwards

.calc_label
if you already set labels with taxatree_label: set this to FALSE to use / avoid overwriting that data (ignores ... if FALSE)

layout
any ggraph layout, default is "tree"

layout_seed
any numeric, required if a stochastic igraph layout is named

circular
should the layout be circular?

node_sort
sort nodes by "increasing" or "decreasing" size? NULL for no sorting. Use tax_sort() before taxatree_plots() for finer control.

add_circles
add faint concentric circles to plot, behind each rank?

drop_ranks
drop ranks from tree if not included in stats dataframe

Value

ggplot

Examples

# see taxatree_plots() examples