Calculate distances between pairs of samples in phyloseq object

Can compute various sample-sample distances using the microbiota composition of your samples:

bray curtis ('bray') or any other ecological distance from phyloseq::distance() / vegan::vegdist()

unifrac distances (using the GUniFrac package)

generalised: 'gunifrac' (optionally set weighting alpha in gunifrac alpha)

unweighted: 'unifrac'

weighted: 'wunifrac'

variance adjusted weighted: 'va-wunifrac'

aitchison distance (euclidean distance after centered log ratio transform clr, see details)

euclidean distance

Use dist_calc with ps_extra output of tax_transform (or tax_agg). It returns a ps_extra object containing the phyloseq and the name of the distance used in addition to the distance matrix itself. The resulting object is intended to be piped into ord_calc or dist_permanova functions. Alternatively you can directly access the distance matrix with dist_get().

dist_calc(
  data,
  dist = c("bray", "gunifrac", "unifrac", "wunifrac", "va-wunifrac", "aitchison",
    "euclidean")[1],
  gunifrac_alpha = 0.5,
  ...
)

Arguments

data
ps_extra object output from tax_transform()

dist
name of distance to calculate between pairs of samples

gunifrac_alpha
setting alpha value only relevant if gunifrac distance used

...
optional distance-specific named arguments passed to phyloseq::distance()

Value

list with distance matrix, phyloseq object, and name of distance used

Details

Aitchison distance note: You should EITHER:

skip the dist_calc function and call ord_calc(method = "PCA") directly on an object with taxa transformed with tax_transform(transformation = "clr")

pass an object with untransformed (or 'identity' transformed) taxa to the data argument of dist_calc() and specify dist = "aitchison".

If ordination plots with taxon loading distances are desired, users require option 1. If the distance matrix is required for permanova, users require option 2.

See also

Examples

# bray curtis distance on genera-level features
data("dietswap", package = "microbiome")
bc <- dietswap %>%
  tax_agg("Genus") %>%
  dist_calc("bray")
bc

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 3 taxonomic ranks ]
#> 
#> phyloseq info:
#> tax_agg = Genus tax_transform = NA
#> 
#> bray distance matrix of size 222 
#> 0.7639533 0.7851213 0.6680796 0.7699252 0.80507 ...

class(bc)

#> [1] "ps_extra" "list"    

# gunifrac distance using phyloseq input
data("esophagus", package = "phyloseq")
gunifrac <- esophagus %>%
  dist_calc("gunifrac") %>%
  dist_get()
class(gunifrac)

#> [1] "dist"