Interactively explore microbial compositions of ordinated samples

A Shiny app used like an interactive version of ord_plot (taking the output of ord_calc). You can select samples on an ordination plot and view their composition with stacked barplots.

Once running: click and drag to select 2 or more samples to view their compositions. You can style the ordination plot points and compositional barplot using the options in the left panel.

ord_explore(
  data,
  sample_id = NULL,
  seriate_method = "OLO_ward",
  tax_transform_for_ordering = "identity",
  app_options = list(launch.browser = TRUE),
  ...
)

Arguments

data
ps_extra list output of ord_calc

sample_id
name of sample ID variable to use as default for selecting samples

seriate_method
seriation method to order phyloseq samples by similarity

tax_transform_for_ordering
transform taxa before ordering with ps_seriate

app_options
passed to shinyApp() options argument

...
additional arguments passed to ord_plot

Value

nothing, opens default browser

Details

If you get an interactive error like the one below:

"ids don't have the same length than str (most often, it occurs because of clipping)"

make your points smaller

set point shape to fixed, or to a variable with fewer categories

Limitation: When a selection grouping variable is NA for some samples, that grouping variable cannot be used to select those samples

Examples

library(phyloseq)
library(dplyr)
data("enterotype")

# simple example #
taxa_names(enterotype)[1] <- "unclassified" # replaces the "-1" taxon name
ps <- tax_fix(enterotype) # remove NA taxa

#> Row named: unclassified
#> contains no non-unknown values, returning:
#> 'unclassified' for all replaced levels.
#> Consider editing this tax_table entry manually.

#> Row named: Bacteria
#> contains no non-unknown values, returning:
#> 'Bacteria' for all replaced levels.
#> Consider editing this tax_table entry manually.

ord1 <- ps %>%
  dist_calc("bray") %>%
  ord_calc(method = "PCoA")

# ord_explore(data = ord1, auto_caption = NA)

# constrained ordination example #

data("dietswap", package = "microbiome")

# create a couple of numerical variables to use as constraints
dietswap <- dietswap %>%
  ps_mutate(
    weight = recode(bmi_group, obese = 3, overweight = 2, lean = 1),
    female = if_else(sex == "female", true = 1, false = 0)
  ) %>%
  tax_agg("Genus")

constrained_aitchison_rda <- dietswap %>%
  tax_transform("clr") %>%
  ord_calc(constraints = c("weight", "female"))

#> 
#> Centering (mean) and scaling (sd) the constraints and conditions: 

#> 	weight

#> 	female

# constrained_aitchison_rda %>%
#   ord_explore(
#     plot_taxa = 1:5, tax_lab_style = list(size = 3),
#     constraint_lab_style = list(size = 4), auto_caption = 7
#   )
# try changing the point colour to bmi_group or similar
# (style interactively! e.g. colour doesn't work as argument to ord_explore)

# another dataset
# microbiomeutilities::hmp2 %>%
#   tax_fix() %>%
#   dist_calc("aitchison") %>%
#   ord_calc() %>%
#   ord_explore()

# another dataset
data("soilrep", package = "phyloseq")
# test auto creation of SAMPLE var
ps <- soilrep %>% ps_select(-Sample)
# The barplot is actually quite useless with the 16000+ anonymous OTUs
# in this dataset, but the 1000s of unmerged "other" categories do render
# phyloseq_validate(ps) %>%
#   tax_fix() %>%
#   dist_calc("aitchison") %>%
#   ord_calc() %>%
#   ord_explore()