Interactively explore microbial compositions of ordinated samples

A Shiny app used like an interactive version of ord_plot(). You can select samples on an ordination plot and view their composition with stacked barplots.

The ord_explore() data argument takes either:

the output of ord_calc() (i.e. a ps_extra with an ordination)

a plain phyloseq object: ord_explore() will help you build an ordination

Once the app is running (in your browser):

edit the ordination if required

style the ordination plot (e.g. choose dimensions; set colour and size; ...)

click or use the lasso tool to select 2 or more samples to view their compositions

style the taxonomic compositions barplot

stop the app by clicking the red stop button in the R console (closing the web browser window doesn't stop the app)

ord_explore(
  data,
  sample_id = NULL,
  seriate_method = "OLO_ward",
  tax_transform_for_ordering = "identity",
  app_options = list(launch.browser = TRUE),
  plot_widths = c(7, 9),
  ...
)

Arguments

data
ps_extra list output of ord_calc, or phyloseq

sample_id
name of sample ID variable to use as default for selecting samples

seriate_method
seriation method to order phyloseq samples by similarity

tax_transform_for_ordering
transform taxa before ordering with ps_seriate

app_options
passed to shinyApp() options argument

plot_widths
widths of plots in inches, including any legends (first number is ordination, second is composition barplot)

...
additional arguments passed to ord_plot

Value

nothing, opens default browser

Details

If you get an interactive error like the one below:

"ids don't have the same length than str (most often, it occurs because of clipping)"

make your points smaller

set point shape to fixed, or to a variable with fewer categories

Limitation: When a selection grouping variable is NA for some samples, that grouping variable cannot be used to select those samples

Examples

# example code only runs in interactive R session
if (interactive()) {
  library(phyloseq)
  library(dplyr)

  # example of quickstart approach with interactive ordination calculation #
  corncob::ibd_phylo %>%
    tax_fix() %>%
    ord_explore()

  # simple example with precalculated ordination #
  data("enterotype")
  taxa_names(enterotype)[1] <- "unclassified" # replaces the "-1" taxon name
  ps <- tax_fix(enterotype) # remove NA taxa
  ord1 <- ps %>%
    tax_transform("identity", rank = "Genus") %>%
    dist_calc("bray") %>%
    ord_calc(method = "PCoA")

  ord_explore(data = ord1, auto_caption = 6)

  # constrained ordination example #
  data("dietswap", package = "microbiome")

  # create a couple of numerical variables to use as constraints
  dietswap <- dietswap %>%
    ps_mutate(
      weight = recode(bmi_group, obese = 3, overweight = 2, lean = 1),
      female = if_else(sex == "female", true = 1, false = 0)
    ) %>%
    tax_agg("Genus")

  constrained_aitchison_rda <- dietswap %>%
    tax_transform("clr") %>%
    ord_calc(constraints = c("weight", "female"))

  # label style arguments can be passed to ord_explore
  constrained_aitchison_rda %>%
    ord_explore(
      tax_lab_style = list(size = 3),
      constraint_lab_style = list(size = 4), auto_caption = 6
    )
  # Try changing the point colour to bmi_group or similar
  # Style points interactively!
  # (setting colour/shape/etc as arguments doesn't work)

  # Another dataset, where "size" variable drives gradient on PC1
  # Try setting size and/or alpha to correspond to "size"!
  # Then edit the ordination to use "size" as a condition, see what happens
  # hmp2 <- microbiomeutilities::hmp2
  hmp2 %>%
    tax_fix() %>%
    tax_transform(rank = "Genus", "identity") %>%
    dist_calc("aitchison") %>%
    ord_calc() %>%
    ord_explore()

  # another dataset
  data("soilrep", package = "phyloseq")
  # test auto creation of SAMPLE var
  ps <- soilrep %>% ps_select(-Sample)
  # The barplot is actually quite useless with the 16000+ anonymous OTUs
  # in this dataset, but the 1000s of unmerged "other" categories do render
  phyloseq_validate(ps) %>%
    tax_fix() %>%
    dist_calc("aitchison") %>%
    ord_calc() %>%
    ord_explore()
}