De-duplicate phyloseq samples

Use 1 or more variables in the sample_data to identify and remove duplicate samples (leaving 1 per category).

methods:

method = "readcount" keeps the one sample in each duplicate group with the highest total number of reads according to sample_sums

method = "first" keeps the first sample in each duplicate group encountered in the row order of the sample_data

method = "last" keeps the last sample in each duplicate group encountered in the row order of the sample_data

ps_dedupe(ps, vars, method = "readcount", verbose = TRUE)

Arguments

ps
phyloseq with sample data

vars
names of variables, whose (combined) levels identify groups from which only 1 sample is desired

method
keep 1 sample with max "readcount" or the "first" or "last" samples encountered in given sample_data order for each dupe group

verbose
message names of samples removed if TRUE (or "debug" for more info, when method = "readcount")

Value

phyloseq object

Details

What happens when duplicated samples have exactly equal readcounts in method = "readcount"? The first encountered maximum is kept (in sample_data row order, like method = "first")

See also

Examples

library(phyloseq)
library(microbiome)
data("dietswap")

dietswap

#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 3 taxonomic ranks ]

# let's pretend the dietswap data contains technical replicates from each subject
# we want to keep only one of them
ps_dedupe(dietswap, vars = "subject", method = "readcount", verbose = TRUE)

#> 184 samples being removed:
#> Sample-1; Sample-4; Sample-5; Sample-7; Sample-9; Sample-10; Sample-11; Sample-13; Sample-14; Sample-15; Sample-16; Sample-17; Sample-18; Sample-19; Sample-20; Sample-21; Sample-22; Sample-24; Sample-25; Sample-26; Sample-27; Sample-28; Sample-29; Sample-30; Sample-31; Sample-32; Sample-33; Sample-34; Sample-35; Sample-36; Sample-37; Sample-39; Sample-40; Sample-41; Sample-42; Sample-43; Sample-44; Sample-45; Sample-47; Sample-48; Sample-50; Sample-51; Sample-52; Sample-53; Sample-55; Sample-56; Sample-57; Sample-58; Sample-59; Sample-60; Sample-61; Sample-62; Sample-63; Sample-64; Sample-65; Sample-66; Sample-67; Sample-68; Sample-69; Sample-70; Sample-71; Sample-72; Sample-73; Sample-74; Sample-75; Sample-76; Sample-78; Sample-79; Sample-80; Sample-82; Sample-83; Sample-84; Sample-85; Sample-86; Sample-87; Sample-88; Sample-90; Sample-92; Sample-93; Sample-94; Sample-95; Sample-97; Sample-99; Sample-100; Sample-101; Sample-102; Sample-103; Sample-104; Sample-105; Sample-108; Sample-109; Sample-112; Sample-113; Sample-114; Sample-115; Sample-117; Sample-119; Sample-120; Sample-121; Sample-122; Sample-124; Sample-125; Sample-126; Sample-127; Sample-128; Sample-129; Sample-131; Sample-132; Sample-133; Sample-134; Sample-135; Sample-136; Sample-137; Sample-138; Sample-139; Sample-140; Sample-141; Sample-142; Sample-144; Sample-145; Sample-146; Sample-147; Sample-148; Sample-151; Sample-152; Sample-153; Sample-154; Sample-155; Sample-158; Sample-159; Sample-160; Sample-162; Sample-163; Sample-164; Sample-166; Sample-168; Sample-170; Sample-171; Sample-172; Sample-173; Sample-174; Sample-175; Sample-176; Sample-177; Sample-178; Sample-179; Sample-180; Sample-182; Sample-183; Sample-184; Sample-185; Sample-186; Sample-187; Sample-188; Sample-189; Sample-190; Sample-191; Sample-192; Sample-193; Sample-194; Sample-195; Sample-196; Sample-197; Sample-198; Sample-199; Sample-201; Sample-202; Sample-204; Sample-205; Sample-206; Sample-207; Sample-208; Sample-209; Sample-211; Sample-212; Sample-213; Sample-214; Sample-215; Sample-216; Sample-217; Sample-218; Sample-219; Sample-220; Sample-221

#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 38 samples ]
#> sample_data() Sample Data:       [ 38 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 3 taxonomic ranks ]

# contrived example to show identifying "duplicates" via the interaction of multiple columns
ps1 <- ps_dedupe(dietswap,
  method = "readcount", verbose = TRUE,
  vars = c("timepoint", "group", "bmi_group")
)

#> 204 samples being removed:
#> Sample-1; Sample-4; Sample-5; Sample-6; Sample-7; Sample-8; Sample-9; Sample-10; Sample-13; Sample-14; Sample-15; Sample-16; Sample-18; Sample-19; Sample-21; Sample-22; Sample-23; Sample-24; Sample-26; Sample-27; Sample-29; Sample-30; Sample-31; Sample-32; Sample-33; Sample-34; Sample-35; Sample-36; Sample-37; Sample-39; Sample-40; Sample-41; Sample-42; Sample-43; Sample-44; Sample-45; Sample-47; Sample-48; Sample-49; Sample-50; Sample-51; Sample-52; Sample-53; Sample-55; Sample-56; Sample-57; Sample-58; Sample-59; Sample-60; Sample-61; Sample-62; Sample-63; Sample-64; Sample-65; Sample-66; Sample-67; Sample-68; Sample-69; Sample-70; Sample-71; Sample-72; Sample-73; Sample-74; Sample-75; Sample-76; Sample-78; Sample-79; Sample-80; Sample-81; Sample-82; Sample-83; Sample-84; Sample-85; Sample-86; Sample-87; Sample-88; Sample-89; Sample-90; Sample-91; Sample-92; Sample-93; Sample-94; Sample-95; Sample-96; Sample-97; Sample-98; Sample-99; Sample-100; Sample-102; Sample-103; Sample-104; Sample-105; Sample-106; Sample-107; Sample-108; Sample-109; Sample-110; Sample-111; Sample-112; Sample-113; Sample-114; Sample-115; Sample-116; Sample-117; Sample-118; Sample-119; Sample-120; Sample-121; Sample-122; Sample-123; Sample-124; Sample-125; Sample-126; Sample-127; Sample-128; Sample-129; Sample-130; Sample-131; Sample-132; Sample-133; Sample-134; Sample-135; Sample-136; Sample-137; Sample-138; Sample-139; Sample-140; Sample-141; Sample-142; Sample-144; Sample-145; Sample-146; Sample-147; Sample-148; Sample-149; Sample-150; Sample-151; Sample-152; Sample-153; Sample-154; Sample-155; Sample-157; Sample-158; Sample-159; Sample-160; Sample-161; Sample-162; Sample-163; Sample-164; Sample-165; Sample-166; Sample-167; Sample-168; Sample-169; Sample-170; Sample-171; Sample-172; Sample-173; Sample-174; Sample-175; Sample-176; Sample-177; Sample-178; Sample-179; Sample-180; Sample-181; Sample-182; Sample-183; Sample-184; Sample-185; Sample-186; Sample-187; Sample-188; Sample-189; Sample-190; Sample-191; Sample-192; Sample-193; Sample-194; Sample-195; Sample-196; Sample-197; Sample-198; Sample-199; Sample-200; Sample-201; Sample-202; Sample-203; Sample-204; Sample-205; Sample-206; Sample-207; Sample-209; Sample-210; Sample-211; Sample-212; Sample-213; Sample-215; Sample-216; Sample-218; Sample-219; Sample-220; Sample-221; Sample-222

sample_data(ps1)

#> Sample Data:        [18 samples by 8 sample variables]:
#>            subject    sex nationality group     sample timepoint
#> Sample-2       nms   male         AFR    HE   Sample-2         2
#> Sample-3       olt   male         AFR    HE   Sample-3         2
#> Sample-11      olt   male         AFR    HE  Sample-11         3
#> Sample-12      pku female         AFR    HE  Sample-12         3
#> Sample-17      ufm   male         AFR    HE  Sample-17         3
#> Sample-20      pku female         AFR    DI  Sample-20         4
#> Sample-25      ufm   male         AFR    DI  Sample-25         4
#> Sample-28      pku female         AFR    DI  Sample-28         5
#> Sample-38      nmz   male         AAM    ED  Sample-38         1
#> Sample-46      mni female         AAM    ED  Sample-46         6
#> Sample-54      dwc   male         AFR    ED  Sample-54         6
#> Sample-77      zaq female         AFR    DI  Sample-77         4
#> Sample-101     zaq female         AFR    DI Sample-101         5
#> Sample-143     kpp   male         AAM    DI Sample-143         5
#> Sample-156     gtd   male         AAM    HE Sample-156         2
#> Sample-208     olt   male         AFR    ED Sample-208         1
#> Sample-214     ufm   male         AFR    ED Sample-214         1
#> Sample-217     pku female         AFR    ED Sample-217         6
#>            timepoint.within.group  bmi_group
#> Sample-2                        1       lean
#> Sample-3                        1 overweight
#> Sample-11                       2 overweight
#> Sample-12                       2      obese
#> Sample-17                       2       lean
#> Sample-20                       1      obese
#> Sample-25                       1       lean
#> Sample-28                       2      obese
#> Sample-38                       1      obese
#> Sample-46                       2 overweight
#> Sample-54                       2       lean
#> Sample-77                       1 overweight
#> Sample-101                      2 overweight
#> Sample-143                      2       lean
#> Sample-156                      1      obese
#> Sample-208                      1 overweight
#> Sample-214                      1       lean
#> Sample-217                      2      obese

ps2 <- ps_dedupe(dietswap,
  method = "first", verbose = TRUE,
  vars = c("timepoint", "group", "bmi_group")
)

#> 204 samples being removed:
#> Sample-5; Sample-6; Sample-7; Sample-8; Sample-9; Sample-13; Sample-14; Sample-15; Sample-16; Sample-17; Sample-20; Sample-21; Sample-22; Sample-23; Sample-24; Sample-25; Sample-29; Sample-30; Sample-31; Sample-33; Sample-34; Sample-37; Sample-38; Sample-39; Sample-40; Sample-42; Sample-43; Sample-46; Sample-47; Sample-48; Sample-49; Sample-50; Sample-51; Sample-52; Sample-53; Sample-54; Sample-55; Sample-56; Sample-57; Sample-58; Sample-59; Sample-60; Sample-61; Sample-62; Sample-63; Sample-64; Sample-65; Sample-66; Sample-67; Sample-68; Sample-69; Sample-70; Sample-71; Sample-72; Sample-73; Sample-74; Sample-75; Sample-76; Sample-77; Sample-78; Sample-79; Sample-80; Sample-81; Sample-82; Sample-83; Sample-84; Sample-85; Sample-86; Sample-87; Sample-88; Sample-89; Sample-90; Sample-91; Sample-92; Sample-93; Sample-94; Sample-95; Sample-96; Sample-97; Sample-98; Sample-99; Sample-100; Sample-101; Sample-102; Sample-103; Sample-104; Sample-105; Sample-106; Sample-107; Sample-108; Sample-109; Sample-110; Sample-111; Sample-112; Sample-113; Sample-114; Sample-115; Sample-116; Sample-117; Sample-118; Sample-119; Sample-120; Sample-121; Sample-122; Sample-123; Sample-124; Sample-125; Sample-126; Sample-127; Sample-128; Sample-129; Sample-130; Sample-131; Sample-132; Sample-133; Sample-134; Sample-135; Sample-136; Sample-137; Sample-138; Sample-139; Sample-140; Sample-141; Sample-142; Sample-143; Sample-144; Sample-145; Sample-146; Sample-147; Sample-148; Sample-149; Sample-150; Sample-151; Sample-152; Sample-153; Sample-154; Sample-155; Sample-156; Sample-157; Sample-158; Sample-159; Sample-160; Sample-161; Sample-162; Sample-163; Sample-164; Sample-165; Sample-166; Sample-167; Sample-168; Sample-169; Sample-170; Sample-171; Sample-172; Sample-173; Sample-174; Sample-175; Sample-176; Sample-177; Sample-178; Sample-179; Sample-180; Sample-181; Sample-182; Sample-183; Sample-184; Sample-185; Sample-186; Sample-187; Sample-188; Sample-189; Sample-190; Sample-191; Sample-192; Sample-193; Sample-194; Sample-195; Sample-196; Sample-197; Sample-198; Sample-199; Sample-200; Sample-201; Sample-202; Sample-203; Sample-204; Sample-205; Sample-206; Sample-207; Sample-208; Sample-209; Sample-210; Sample-211; Sample-212; Sample-213; Sample-214; Sample-215; Sample-216; Sample-217; Sample-218; Sample-219; Sample-220; Sample-221; Sample-222

sample_data(ps2)

#> Sample Data:        [18 samples by 8 sample variables]:
#>           subject    sex nationality group    sample timepoint
#> Sample-1      byn   male         AAM    DI  Sample-1         4
#> Sample-2      nms   male         AFR    HE  Sample-2         2
#> Sample-3      olt   male         AFR    HE  Sample-3         2
#> Sample-4      pku female         AFR    HE  Sample-4         2
#> Sample-10     nms   male         AFR    HE Sample-10         3
#> Sample-11     olt   male         AFR    HE Sample-11         3
#> Sample-12     pku female         AFR    HE Sample-12         3
#> Sample-18     nms   male         AFR    DI Sample-18         4
#> Sample-19     olt   male         AFR    DI Sample-19         4
#> Sample-26     nms   male         AFR    DI Sample-26         5
#> Sample-27     olt   male         AFR    DI Sample-27         5
#> Sample-28     pku female         AFR    DI Sample-28         5
#> Sample-32     azh female         AAM    ED Sample-32         1
#> Sample-35     kpp   male         AAM    ED Sample-35         1
#> Sample-36     lot female         AAM    ED Sample-36         1
#> Sample-41     azh female         AAM    ED Sample-41         6
#> Sample-44     kpp   male         AAM    ED Sample-44         6
#> Sample-45     lot female         AAM    ED Sample-45         6
#>           timepoint.within.group  bmi_group
#> Sample-1                       1      obese
#> Sample-2                       1       lean
#> Sample-3                       1 overweight
#> Sample-4                       1      obese
#> Sample-10                      2       lean
#> Sample-11                      2 overweight
#> Sample-12                      2      obese
#> Sample-18                      1       lean
#> Sample-19                      1 overweight
#> Sample-26                      2       lean
#> Sample-27                      2 overweight
#> Sample-28                      2      obese
#> Sample-32                      1 overweight
#> Sample-35                      1       lean
#> Sample-36                      1      obese
#> Sample-41                      2 overweight
#> Sample-44                      2       lean
#> Sample-45                      2      obese