Aggregate taxa and create ps_extra

tax_agg sums the abundances of the phyloseq taxa at the given rank. It records the tax_agg rank argument in the info of the ps_extra object output. This ps_extra object tracks aggregation, and any further transformations and scaling, to help you keep track of what you have done with your phyloseq object. tax_agg allows you to pass NA or "unique" to the rank argument which will NOT aggregate the taxa. It always adds a "unique" rank, which matches the tax_table row names (after any aggregation performed).

You should NOT use the top_N argument yourself unless you know what you are doing. top_N provides a feature similar to the deprecated microbiome function aggregate_top_taxa which is primarily useful for decluttering compositional barplots. microViz comp_barplot (and ord_plot_iris) already run tax_agg with a top_N argument for you, so you should not. The tax_table produced when using top_N is otherwise INVALID FOR MOST OTHER ANALYSES.

tax_agg(ps, rank = "unique", sort_by = NA, top_N = NA)

Arguments

ps
phyloseq object

rank
NA or "unique" or name of valid taxonomic rank (try phyloseq::rank_names(ps))

sort_by
if not NA, how should the taxa be sorted, uses tax_sort(), takes same options as by arg

top_N
NA does nothing, but if top_N is a number, it creates an extra tax_table column called top, which is the same as the unique column for the first top_N number of taxa, and "other" otherwise.

Value

ps_extra list object including phyloseq and tax_agg rank info

Details

This function is inspired by microbiome::aggregate_taxa. However if microbiome::aggregate_taxa is used, microViz cannot track this aggregation. Except for the ordering of taxa, the resulting phyloseq objects are identical for aggregating a phyloseq with n ranks up to the n-1 rank. Aggregating at rank n with tax_agg still generates a "unique" column, but aggregate_taxa does not do this.

Examples

library(microbiome)
data("dietswap", package = "microbiome")

tax_agg(ps = dietswap, "Phylum") %>%
  ps_get() %>%
  tax_table()

#> Taxonomy Table:     [8 taxa by 2 taxonomic ranks]:
#>                 Phylum            unique           
#> Actinobacteria  "Actinobacteria"  "Actinobacteria" 
#> Firmicutes      "Firmicutes"      "Firmicutes"     
#> Proteobacteria  "Proteobacteria"  "Proteobacteria" 
#> Verrucomicrobia "Verrucomicrobia" "Verrucomicrobia"
#> Bacteroidetes   "Bacteroidetes"   "Bacteroidetes"  
#> Spirochaetes    "Spirochaetes"    "Spirochaetes"   
#> Fusobacteria    "Fusobacteria"    "Fusobacteria"   
#> Cyanobacteria   "Cyanobacteria"   "Cyanobacteria"  

tax_agg(ps = dietswap, "Family") %>%
  ps_get() %>%
  tax_table()

#> Taxonomy Table:     [22 taxa by 3 taxonomic ranks]:
#>                           Phylum            Family                     
#> Actinobacteria            "Actinobacteria"  "Actinobacteria"           
#> Bacilli                   "Firmicutes"      "Bacilli"                  
#> Proteobacteria            "Proteobacteria"  "Proteobacteria"           
#> Verrucomicrobia           "Verrucomicrobia" "Verrucomicrobia"          
#> Bacteroidetes             "Bacteroidetes"   "Bacteroidetes"            
#> Clostridium cluster XV    "Firmicutes"      "Clostridium cluster XV"   
#> Clostridium cluster XIVa  "Firmicutes"      "Clostridium cluster XIVa" 
#> Clostridium cluster IV    "Firmicutes"      "Clostridium cluster IV"   
#> Clostridium cluster XI    "Firmicutes"      "Clostridium cluster XI"   
#> Asteroleplasma            "Firmicutes"      "Asteroleplasma"           
#> Spirochaetes              "Spirochaetes"    "Spirochaetes"             
#> Clostridium cluster XVI   "Firmicutes"      "Clostridium cluster XVI"  
#> Clostridium cluster XVII  "Firmicutes"      "Clostridium cluster XVII" 
#> Clostridium cluster I     "Firmicutes"      "Clostridium cluster I"    
#> Clostridium cluster XVIII "Firmicutes"      "Clostridium cluster XVIII"
#> Clostridium cluster III   "Firmicutes"      "Clostridium cluster III"  
#> Clostridium cluster IX    "Firmicutes"      "Clostridium cluster IX"   
#> Fusobacteria              "Fusobacteria"    "Fusobacteria"             
#> Clostridium cluster XIII  "Firmicutes"      "Clostridium cluster XIII" 
#> Cyanobacteria             "Cyanobacteria"   "Cyanobacteria"            
#> Uncultured Clostridiales  "Firmicutes"      "Uncultured Clostridiales" 
#> Uncultured Mollicutes     "Firmicutes"      "Uncultured Mollicutes"    
#>                           unique                     
#> Actinobacteria            "Actinobacteria"           
#> Bacilli                   "Bacilli"                  
#> Proteobacteria            "Proteobacteria"           
#> Verrucomicrobia           "Verrucomicrobia"          
#> Bacteroidetes             "Bacteroidetes"            
#> Clostridium cluster XV    "Clostridium cluster XV"   
#> Clostridium cluster XIVa  "Clostridium cluster XIVa" 
#> Clostridium cluster IV    "Clostridium cluster IV"   
#> Clostridium cluster XI    "Clostridium cluster XI"   
#> Asteroleplasma            "Asteroleplasma"           
#> Spirochaetes              "Spirochaetes"             
#> Clostridium cluster XVI   "Clostridium cluster XVI"  
#> Clostridium cluster XVII  "Clostridium cluster XVII" 
#> Clostridium cluster I     "Clostridium cluster I"    
#> Clostridium cluster XVIII "Clostridium cluster XVIII"
#> Clostridium cluster III   "Clostridium cluster III"  
#> Clostridium cluster IX    "Clostridium cluster IX"   
#> Fusobacteria              "Fusobacteria"             
#> Clostridium cluster XIII  "Clostridium cluster XIII" 
#> Cyanobacteria             "Cyanobacteria"            
#> Uncultured Clostridiales  "Uncultured Clostridiales" 
#> Uncultured Mollicutes     "Uncultured Mollicutes"    

# this won't aggregate taxa,
# but just add a new rank called unique, equal to taxa_names
tax_agg(ps = dietswap, NA)

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 4 taxonomic ranks ]
#> 
#> phyloseq info:
#> tax_agg = unique tax_transform = NA

identical(tax_agg(dietswap, NA), tax_agg(dietswap, "unique")) # TRUE

#> [1] TRUE

# create some missing values
tax_table(dietswap)[3:7, "Genus"] <- "g__"
# this will produce an error, instructing the user to use fill_unknowns
# tax_agg(ps = dietswap, "Genus")
# this will then work:
dietswap %>%
  tax_fill_unknowns() %>%
  tax_agg("Genus")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 128 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 128 taxa by 4 taxonomic ranks ]
#> 
#> phyloseq info:
#> tax_agg = Genus tax_transform = NA

# you can replace long unknown values
tax_table(dietswap)[13:17, "Family"] <- "some_unknown_family"
# default tax_fill_unknowns settings won't catch it (too long)
dietswap %>%
  tax_fill_unknowns(unknowns = "some_unknown_family") %>%
  tax_agg("Family")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 25 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 25 taxa by 3 taxonomic ranks ]
#> 
#> phyloseq info:
#> tax_agg = Family tax_transform = NA