Mean-center and SD-scale taxa in phyloseq

Wrapper for applying base scale function to phyloseq otu_table

tax_scale(data, center = TRUE, scale = TRUE, do = NA)

Arguments

data
phyloseq or ps_extra or otu_table

center
if TRUE: center each taxon by subtracting its mean

scale
if TRUE, divide each centred taxon by its standard deviation (or divide by RMS if not centred!)

do
alternative argument that overrides center and scale options! takes "both", "scale", "center" or "neither"

Examples

library(phyloseq)
library(microbiome)
data("dietswap")
ps <- dietswap
ps %>%
  otu_get() %>%
  .[1:6, 1:6]

#> OTU Table:          [6 taxa and 6 samples]
#>                      taxa are columns
#>          Actinomycetaceae Aerococcus Aeromonas Akkermansia
#> Sample-1                0          0         0          18
#> Sample-2                1          0         0          97
#> Sample-3                0          0         0          67
#> Sample-4                1          0         0         256
#> Sample-5                0          0         0          21
#> Sample-6                0          0         0          16
#>          Alcaligenes faecalis et rel. Allistipes et rel.
#> Sample-1                            1                336
#> Sample-2                            2                 63
#> Sample-3                            3                 36
#> Sample-4                            2                 96
#> Sample-5                            2                 49
#> Sample-6                            2                 17

# standard use (mean center and SD scale)
tax_scale(ps) %>%
  otu_get() %>%
  .[1:6, 1:6]

#> OTU Table:          [6 taxa and 6 samples]
#>                      taxa are columns
#>          Actinomycetaceae Aerococcus   Aeromonas Akkermansia
#> Sample-1       -0.2780648        NaN -0.09513127  -0.3332126
#> Sample-2        3.5800844        NaN -0.09513127  -0.0622108
#> Sample-3       -0.2780648        NaN -0.09513127  -0.1651229
#> Sample-4        3.5800844        NaN -0.09513127   0.4832231
#> Sample-5       -0.2780648        NaN -0.09513127  -0.3229214
#> Sample-6       -0.2780648        NaN -0.09513127  -0.3400734
#>          Alcaligenes faecalis et rel. Allistipes et rel.
#> Sample-1                   -0.3262211          0.2690184
#> Sample-2                   -0.2720137         -0.5929499
#> Sample-3                   -0.2178063         -0.6781996
#> Sample-4                   -0.2720137         -0.4887560
#> Sample-5                   -0.2720137         -0.6371534
#> Sample-6                   -0.2720137         -0.7381900

# RMS scale only (directly on otu_table)
otu_table(ps) %>%
  tax_scale(center = FALSE) %>%
  .[1:6, 1:6] # Aerococcus is NaN as prevalence=0

#> OTU Table:          [6 taxa and 6 samples]
#>                      taxa are columns
#>          Actinomycetaceae Aerococcus Aeromonas Akkermansia
#> Sample-1         0.000000        NaN         0  0.05741261
#> Sample-2         3.716517        NaN         0  0.30939020
#> Sample-3         0.000000        NaN         0  0.21370251
#> Sample-4         3.716517        NaN         0  0.81653496
#> Sample-5         0.000000        NaN         0  0.06698138
#> Sample-6         0.000000        NaN         0  0.05103343
#>          Alcaligenes faecalis et rel. Allistipes et rel.
#> Sample-1                   0.05065049         0.83097699
#> Sample-2                   0.10130097         0.15580819
#> Sample-3                   0.15195146         0.08903325
#> Sample-4                   0.10130097         0.23742200
#> Sample-5                   0.10130097         0.12118414
#> Sample-6                   0.10130097         0.04204348

# example using alternative `do` argument (to center only, no scaling)
tax_scale(ps, do = "center") %>%
  otu_get() %>%
  .[1:6, 1:6]

#> OTU Table:          [6 taxa and 6 samples]
#>                      taxa are columns
#>          Actinomycetaceae Aerococcus    Aeromonas Akkermansia
#> Sample-1      -0.07207207          0 -0.009009009   -97.13514
#> Sample-2       0.92792793          0 -0.009009009   -18.13514
#> Sample-3      -0.07207207          0 -0.009009009   -48.13514
#> Sample-4       0.92792793          0 -0.009009009   140.86486
#> Sample-5      -0.07207207          0 -0.009009009   -94.13514
#> Sample-6      -0.07207207          0 -0.009009009   -99.13514
#>          Alcaligenes faecalis et rel. Allistipes et rel.
#> Sample-1                    -6.018018            85.2027
#> Sample-2                    -5.018018          -187.7973
#> Sample-3                    -4.018018          -214.7973
#> Sample-4                    -5.018018          -154.7973
#> Sample-5                    -5.018018          -201.7973
#> Sample-6                    -5.018018          -233.7973