Transform taxa in phyloseq object and record transformation

Transform taxa features, and optionally aggregate at specified taxonomic rank beforehand. You can pipe the results of tax_agg into tax_transform, or equivalently set the rank argument in tax_transform.

tax_transform(
  data,
  trans,
  rank = NA,
  keep_counts = TRUE,
  chain = FALSE,
  zero_replace = 0,
  transformation = NULL,
  ...
)

Arguments

data
a phyloseq object or ps_extra list output from tax_agg

trans
any valid taxa transformation (e.g. from microbiome::transform)

rank
If data is phyloseq: data are aggregated at this rank before transforming. If NA, runs tax_agg(data, rank = NA). If rank is NA and data is already ps_extra, any preceding aggregation is left as is.

keep_counts
if TRUE, store the pre-transformation count data in ps_extra counts slot

chain
if TRUE, transforming again is possible when data are already transformed i.e. multiple transformations can be chained with multiple tax_transform calls

zero_replace
Replace any zeros with this value before transforming. Either a numeric, or "halfmin" which replaces zeros with half of the smallest value across the entire dataset. Beware: the choice of zero replacement is not tracked in the ps_extra output.

transformation
deprecated, use trans instead!

...
any extra arguments passed to microbiome::transform or pass undetected = a number when using trans = "binary"

Value

ps_extra list including phyloseq object and info

Details

This function often uses microbiome::transform internally and can perform the same transformations, including many from vegan::decostand (where the default MARGIN = 2). See below for notes about some of the available transformations.

tax_transform returns a ps_extra list containing the transformed phyloseq object and extra info (used for annotating ord_plot ordinations):

tax_transform (a string recording the transformation),

tax_agg (a string recording the taxonomic aggregation rank if specified here or earlier in tax_agg).

A few commonly used transformations:

"clr" performs the centered log ratio transformation using microbiome::transform

"compositional" converts the data into proportions, from 0 to 1.

"identity" does not transform the data, and records this choice for ord_plot

"binary" can be used to transform tax abundances into presence/abundance data.

"log2" which performs a log base 2 transformation (don't forget to set zero_replace if there are any zeros in your data)

clr transformation note

If any values are zero, the clr transform routine first adds a small pseudocount of min(relative abundance)/2 to all values. To avoid this, you can replace any zeros in advance by setting zero_replace to a number > 0.

Binary transformation notes

By default, otu_table values of 0 are kept as 0, and all positive values are converted to 1 (like decostand(method = "pa")). You can set a different threshold, by passing e.g. undetected = 10, for example, in which case all abundances of 10 or below would be converted to 0. All abundances above 10 would be converted to 1s.

Beware that the choice of detection threshold is not tracked in the ps_extra.

See also

Examples

library(microbiome)
data("dietswap", package = "microbiome")

# aggregate taxa at Phylum level and center log ratio transform the phyla counts
tax_transform(dietswap, trans = "clr", rank = "Phylum")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 8 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 8 taxa by 1 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = Phylum tax_transform = clr
#> 
#> $counts OTU Table: [ 8 taxa and 222 samples ]

# this is equivalent to the two-step method (agg then transform)
tax_agg(dietswap, rank = "Phylum") %>% tax_transform("clr")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 8 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 8 taxa by 1 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = Phylum tax_transform = clr
#> 
#> $counts OTU Table: [ 8 taxa and 222 samples ]

# does nothing except record tax_agg as "unique" and tax_transform as "identity" in ps_extra info
dietswap %>% tax_transform("identity", rank = NA)

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 4 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = unique tax_transform = identity

# binary transformation (convert abundances to presence/absence or detected/undetected)
tax_transform(dietswap, trans = "binary", rank = "Genus")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 3 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = Genus tax_transform = binary
#> 
#> $counts OTU Table: [ 130 taxa and 222 samples ]

# change detection threshold by setting undetected argument (default is 0)
tax_transform(dietswap, trans = "binary", rank = "Genus", undetected = 50) %>%
  otu_get() %>%
  .[1:6, 1:4]

#> OTU Table:          [4 taxa and 6 samples]
#>                      taxa are columns
#>          Actinomycetaceae Aerococcus Aeromonas Akkermansia
#> Sample-1                0          0         0           0
#> Sample-2                0          0         0           1
#> Sample-3                0          0         0           1
#> Sample-4                0          0         0           1
#> Sample-5                0          0         0           0
#> Sample-6                0          0         0           0

# log2 transformation after replacing all zeros with a pseudocount of 1
tax_transform(dietswap, trans = "log2", rank = "Family", zero_replace = 1)

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 22 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 22 taxa by 2 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = Family tax_transform = log2
#> 
#> $counts OTU Table: [ 22 taxa and 222 samples ]

# log2 transformation after replacing all zeros with a pseudocount of half
# the minimum non-zero count value in the aggregated dataset
tax_transform(dietswap, trans = "log2", rank = "Family", zero_replace = "halfmin")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 22 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 22 taxa by 2 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = Family tax_transform = log2
#> 
#> $counts OTU Table: [ 22 taxa and 222 samples ]