Transform taxa in phyloseq object and record transformation

Transform taxa features, and optionally aggregate at specified taxonomic rank beforehand. You can pipe the results of tax_agg into tax_transform, or equivalently set the rank argument in tax_transform.

tax_transform(data, transformation, rank = NA, keep_counts = TRUE, ...)

Arguments

data
ps_extra list output from tax_agg, or a phyloseq object

transformation
any valid taxa transformation (e.g. from microbiome::transform)

rank
If data is phyloseq: data are aggregated at this rank before transforming. If NA, runs tax_agg(data, rank = NA). If rank is NA and data is already ps_extra, any preceding aggregation is left as is.

keep_counts
if TRUE, store the pre-transformation count data in ps_extra counts slot

...
any extra arguments passed to microbiome::transform or pass undetected = a number when using transformation = "binary"

Value

ps_extra list including phyloseq object and info

Details

This function often uses microbiome::transform internally and can perform the same transformations, including many from vegan::decostand (where the default MARGIN = 2). See below for notes about some of the available transformations.

tax_transform returns a ps_extra list containing the transformed phyloseq object and extra info (used for annotating ord_plot ordinations):

tax_transform (a string recording the transformation),

tax_agg (a string recording the taxonomic aggregation rank if specified here or earlier in tax_agg).

A few commonly used transformations:

"clr" performs the centered log ratio transformation using microbiome::transform, which adds a small pseudocount of min(relative abundance)/2.

"compositional" converts the data into proportions, from 0 to 1.

"identity" does not transform the data, and records this choice for ord_plot

"binary" can be used to transform tax abundances into presence/abundance data.

Binary transform notes:

By default, otu_table values of 0 are kept as 0, and all positive values are converted to 1 (like decostand(method = "pa")). You can set a different threshold, by passing e.g. undetected = 10, for example, in which case all abundances of 10 or below would be converted to 0s. All abundances above 10 would be converted to 1s. Use any numeric value.

See also

Examples

library(microbiome)
data("dietswap", package = "microbiome")

# aggregate taxa at Phylum level and perform the center log ratio transform on the phyla counts
tax_transform(dietswap, transformation = "clr", rank = "Phylum")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 8 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 8 taxa by 1 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = Phylum tax_transform = clr
#> 
#> $counts OTU Table: [ 8 taxa and 222 samples ]

# equivalent to old, two-step method
tax_agg(dietswap, rank = "Phylum") %>% tax_transform("clr")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 8 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 8 taxa by 1 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = Phylum tax_transform = clr
#> 
#> $counts OTU Table: [ 8 taxa and 222 samples ]

# do nothing except record tax_agg as "unique" and tax_transform as "identity" in ps_extra info
dietswap %>% tax_transform("identity", rank = NA)

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 4 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = unique tax_transform = identity

# binary transformation (convert abundances to presence/absence or detected/undetected)
tax_transform(dietswap, transformation = "binary")

#> ps_extra object - a list with phyloseq and extras:
#> 
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:         [ 130 taxa and 222 samples ]
#> sample_data() Sample Data:       [ 222 samples by 8 sample variables ]
#> tax_table()   Taxonomy Table:    [ 130 taxa by 4 taxonomic ranks ]
#> 
#> ps_extra info:
#> tax_agg = unique tax_transform = binary
#> 
#> $counts OTU Table: [ 130 taxa and 222 samples ]

# change detection threshold by setting undetected argument (default is 0)
tax_transform(dietswap, transformation = "binary", undetected = 50) %>%
  otu_get() %>%
  .[1:6, 1:4]

#> OTU Table:          [4 taxa and 6 samples]
#>                      taxa are columns
#>          Actinomycetaceae Aerococcus Aeromonas Akkermansia
#> Sample-1                0          0         0           0
#> Sample-2                0          0         0           1
#> Sample-3                0          0         0           1
#> Sample-4                0          0         0           1
#> Sample-5                0          0         0           0
#> Sample-6                0          0         0           0